Adding --dry-run flag, fixing bugs and correctly flame link
