+++ workflow/cmd_exec.go
@@ -1,4 +1,4 @@
-package run
+package workflow
 
 import (
 	"fmt"
